Update accordion & toggle-group to behave properly
Before this, accordion had odd behavior of duplicating the button when
inside of a toggle-group. This is because toggle-group & accordion both
try to move its children into a div inside the template. This is bad -
it causes all sorts of weird stuff when you have nested containers.

The solution is to just don't: if you want to hide/show something, just
dig out the direct descendants using CSS selectors like :scope >
.class_name.